# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Micro-Saas Ropa API
  description: API completa para el sistema de e-commerce de ropa con gestión de productos, pedidos, usuarios y más.
  version: 1.0.0
  contact:
    name: Soporte Técnico
    email: support@microsaas-ropa.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Servidor de desarrollo
  - url: https://api.microsaas-ropa.com/api
    description: Servidor de producción
paths:
...
  /users/{id}:
    put:
      tags:
        - Usuarios
      summary: Actualizar usuario
      description: Actualiza un usuario existente (solo super_admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID del usuario
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                  description: Nombre completo del usuario
                  example: Juan Pérez
                role:
                  type: string
                  enum: [super_admin, admin_basico, customer]
                  description: Rol del usuario
                  example: admin_basico
                is_active:
                  type: boolean
                  description: Estado del usuario
                  example: true
      responses:
        '200':
          description: Usuario actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /orders:
    get:
      tags:
        - Pedidos
      summary: Listar pedidos
      description: Obtiene una lista de pedidos (requiere autenticación)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Número de página
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Límite de pedidos por página
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          description: Filtrar por estado
          schema:
            type: string
            enum: [pendiente, confirmado, enviado, entregado, cancelado]
      responses:
        '200':
          description: Lista de pedidos obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Pedidos
      summary: Crear pedido
      description: Crea un nuevo pedido (requiere autenticación)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderInput'
      responses:
        '201':
          description: Pedido creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID único del usuario
          example: 123e4567-e89b-12d3-a456-426614174000
        email:
          type: string
          format: email
          description: Email del usuario
          example: admin@example.com
        full_name:
          type: string
          description: Nombre completo del usuario
          example: Administrador del Sistema
        role:
          type: string
          enum: [super_admin, admin_basico, customer]
          description: Rol del usuario
          example: admin_basico
        is_active:
          type: boolean
          description: Estado del usuario
          example: true
        created_at:
          type: string
          format: date-time
          description: Fecha de creación
          example: 2024-01-01T00:00:00Z
        updated_at:
          type: string
          format: date-time
          description: Fecha de última actualización
          example: 2024-01-01T00:00:00Z
...
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID único del pedido
          example: 123e4567-e89b-12d3-a456-426614174000
        customer_name:
          type: string
          description: Nombre del cliente
          example: Juan Pérez
        customer_email:
          type: string
          format: email
          description: Email del cliente
          example: juan@example.com
        status:
          type: string
          enum: [pendiente, confirmado, enviado, entregado, cancelado]
          description: Estado del pedido
          example: pendiente
        total:
          type: number
          format: decimal
          description: Total del pedido
          example: 59.98
        payment_method:
          type: string
          enum: [cash, card, transfer, mercado_pago]
          description: Método de pago
          example: card
        shipping_address:
          type: string
          description: Dirección de envío
          example: Calle Falsa 123, Ciudad, País
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
          description: Items del pedido
        created_at:
          type: string
          format: date-time
          description: Fecha de creación
          example: 2024-01-01T00:00:00Z
        updated_at:
          type: string
          format: date-time
          description: Fecha de última actualización
          example: 2024-01-01T00:00:00Z
    OrderItem:
      type: object
      properties:
        product_id:
          type: string
          format: uuid
          description: ID del producto
          example: 123e4567-e89b-12d3-a456-426614174000
        product_name:
          type: string
          description: Nombre del producto
          example: Camiseta Básica
        quantity:
          type: integer
          minimum: 1
          description: Cantidad
          example: 2
        unit_price:
          type: number
          format: decimal
          description: Precio unitario
          example: 29.99
        total:
          type: number
          format: decimal
          description: Total del item
          example: 59.98
    OrderInput:
      type: object
      required:
        - customer_name
        - customer_email
        - items
        - shipping_address
        - payment_method
      properties:
        customer_name:
          type: string
          minLength: 1
          maxLength: 255
          description: Nombre del cliente
          example: Juan Pérez
        customer_email:
          type: string
          format: email
          description: Email del cliente
          example: juan@example.com
        items:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - product_id
              - quantity
            properties:
              product_id:
                type: string
                format: uuid
                description: ID del producto
                example: 123e4567-e89b-12d3-a456-426614174000
              quantity:
                type: integer
                minimum: 1
                description: Cantidad
                example: 2
        shipping_address:
          type: string
          minLength: 10
          description: Dirección de envío
          example: Calle Falsa 123, Ciudad, País
        payment_method:
          type: string
          enum: [cash, card, transfer, mercado_pago]
          description: Método de pago
          example: card
  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Mensaje de error
                example: Datos inválidos
              details:
                type: string
                description: Detalles adicionales del error
                example: El campo email es requerido
    Unauthorized:
      description: No autorizado
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Mensaje de error
                example: No autorizado
    Forbidden:
      description: Acceso prohibido
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Mensaje de error
                example: Acceso prohibido
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Mensaje de error
                example: Recurso no encontrado
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Autenticación JWT mediante token en cookies o header Authorization